+++ docs/changelog.rst
@@ -9,6 +9,10 @@ CHANGELOG
 
 - Fix access to pictures generated with watermark (#2840)
 
+**New features**
+
+- Allow to filter cirkwi circtuis.xml and pois.xml with portals and structures
+
 
 2.73.0 (2021-12-13)
 -----------------------