+++ docs/changelog.rst
@@ -8,6 +8,7 @@ CHANGELOG
 **Bug fixes**
 
 - Fix access to pictures generated with watermark (#2840)
+- Fix intervention creation and update is now scrollable after merging tabs (#2712)
 
 
 2.73.0 (2021-12-13)