+++ docs/changelog.rst
@@ -9,10 +9,12 @@ CHANGELOG
 
 - Fix access to pictures generated with watermark (#2840)
 - Fix intervention creation and update is now scrollable after merging tabs (#2712)
+- Fix restricted area and restricted area type filters on intervention (#2766)
 
 **New features**
 
 - Allow to filter cirkwi circtuis.xml and pois.xml with portals and structures
+- Add restricted area and restricted area type filters on projects
 
 
 2.73.0 (2021-12-13)