+++ docs/changelog.rst
@@ -8,7 +8,7 @@ CHANGELOG
 **Minor improvements**
 
 - Show every paths in intervention's csv
-- Unshow signage/blade dropdown-toggle with settings `BLADE_ENABLED=False`
+- Hide signage/blade dropdown-toggle with settings `BLADE_ENABLED=False`
 - Remove urls blade with settings `BLADE_ENABLED=False`
 
 **Bug fixes**